Move auth to useAuth + hooks → composables
