+++ apps/frontend/src/api/onRequestFulfilled.ts
@@ -1,4 +1,4 @@
-import { useAuth } from '@/stores/auth';
+import { useAuth } from '@/composables/useAuth';
 import type { InternalAxiosRequestConfig } from 'axios';
 import requestCaseMiddleware from './requestCaseMiddleware';
 import { cloneDeep } from 'lodash-es';
